restructure approaches
