+++ modules/text/module_text_llm/module_text_llm/__main__.py
@@ -1,20 +1,16 @@
-import json
 import os
 from typing import List, Any
 
 import nltk
 import tiktoken
-from module_text_llm.approach_controller import generate
 from athena import app, submission_selector, submissions_consumer, feedback_consumer, feedback_provider, evaluation_provider
 from athena.text import Exercise, Submission, Feedback
 from athena.logger import logger
 
 from module_text_llm.config import Configuration
 from module_text_llm.evaluation import get_feedback_statistics, get_llm_statistics
-from module_text_llm.generate_suggestions import generate_suggestions
 from module_text_llm.generate_evaluation import generate_evaluation
-from module_text_llm.generate_cot_suggestions import generate_cot_suggestions
-
+from module_text_llm.approaches.approach_controller import generate_suggestions
 
 @submissions_consumer
 def receive_submissions(exercise: Exercise, submissions: List[Submission]):
@@ -36,7 +32,7 @@ def process_incoming_feedback(exercise: Exercise, submission: Submission, feedba
 async def suggest_feedback(exercise: Exercise, submission: Submission, is_graded: bool, module_config: Configuration) -> List[Feedback]:
     logger.info("suggest_feedback: %s suggestions for submission %d of exercise %d were requested",
                 "Graded" if is_graded else "Non-graded", submission.id, exercise.id)
-    return await generate(exercise, submission, module_config.approach, module_config.debug)
+    return await generate_suggestions(exercise, submission, module_config.approach, module_config.debug)
 
 
 @evaluation_provider
