+++ modules/text/module_text_llm/module_text_llm/approaches/basic_approach/prompts/generate_suggestions.py
