+++ modules/text/module_text_llm/module_text_llm/approaches/chain_of_thought_approach/prompts/cot_suggestions.py
@@ -18,6 +18,7 @@ The maximal amount of points for this exercise is {max_points}.
 # Grading Instructions
 {grading_instructions}
 
+Respond in json
 """
 
 human_message = """\
